backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-decouple==3.8
Pillow==10.2.0
django-filter==23.5



# === TESTS (à ajouter progressivement) ===
pytest==7.4.4
pytest-django==4.7.0
# pytest-cov==4.1.0
# factory-boy==3.3.0